requirements.txt
pandas==2.0.3
openpyxl==3.1.2
openai==0.28.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
tiktoken==0.5.1
requests==2.31.0
pyairtable==2.2.0
numpy==1.24.3